third_party/blink/web_tests/platform/generic/fast/dom/importNode-unsupported-node-type-expected.txt

importNode should throw informative errors for unsupported node types

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.importNode(doc) threw exception NotSupportedError: Failed to execute 'importNode' on 'Document': The node provided is a document, which may not be imported..
PASS document.importNode(shadow) threw exception NotSupportedError: Failed to execute 'importNode' on 'Document': The node provided is a shadow root, which may not be imported..
PASS successfullyParsed is true

TEST COMPLETE
